Extract libraries from StatBlockLibrary.ts
